+++ CHANGELOG.md
@@ -4,6 +4,10 @@
    Can be edited on `SuperUser` - > `VDBs` page.
    This JSON string will store `Retrieval Filters` for certain `group_vdbs` row.
 
+2. `Query Documents` -> `Queries` page:
+   added `Retrieval Filters`: table of current values + reset button + modal to edit values.
+   This is shown only if `group_vdbs`.`gvdbs_retr_filters` is filled.
+
 ## [0.19a] (2026-01-25)
 
 1. Added new columns: